feat: big optimized jwt

+++ src/users/users.controller.ts
@@ -6,15 +6,16 @@ import {
   Body,
   Param,
   UseGuards,
+  ForbiddenException,
 } from '@nestjs/common';
 import { ApiTags, ApiOperation, ApiBearerAuth } from '@nestjs/swagger';
 import { UsersService } from './users.service';
 import { CreateUserDto } from './dto/create-user.dto';
 import { JwtAuthGuard } from '../auth/jwt-auth.guard';
+import { Authorized } from 'src/auth/decorators/authorized.decorator';
 
 @ApiTags('Users')
 @ApiBearerAuth('access-token')
-@UseGuards(JwtAuthGuard)
 @Controller('users')
 export class UsersController {
   constructor(private readonly usersService: UsersService) {}
@@ -38,8 +39,11 @@ export class UsersController {
   }
 
   @Delete(':id')
-  @ApiOperation({ summary: 'Удалить пользователя' })
-  remove(@Param('id') id: string) {
+  @ApiOperation({ summary: 'Удалить пользователя (только ADMIN)' })
+  remove(@Param('id') id: string, @Authorized() user: any) {
+    if (user.role !== 'ADMIN') {
+      throw new ForbiddenException('Нет прав на удаление пользователя');
+    }
     return this.usersService.remove(+id);
   }
 }
